using size_t

+++ code/src/newneunet/newneunet.c
@@ -381,13 +381,13 @@ float* calculate_output (const neuralnet_t* net, const float* input)
             net->neurons_per_layer[0] * sizeof (float));
     sigmoidize_inplace (current_result_1, net->neurons_per_layer[0]);
 
-    for (uint32_t gap = 0; gap < net->layer_count - 1; ++gap)
+    for (size_t gap = 0; gap < net->layer_count - 1; ++gap)
     {
-        for (uint32_t to = 0; to < net->neurons_per_layer[gap + 1]; ++to)
+        for (size_t to = 0; to < net->neurons_per_layer[gap + 1]; ++to)
         {
             current_result_2[to] = 0.0f;
 
-            for (uint32_t from = 0; from < net->neurons_per_layer[gap]; ++from)
+            for (size_t from = 0; from < net->neurons_per_layer[gap]; ++from)
             {
                 current_result_2[to] +=
                     current_result_1[from] * net->edges[gap][from][to];
